refactor[#287] : 공용 API 를 통해 export 하도록 수정
