+++ src/features/setting/api/index.ts
@@ -1,1 +1,3 @@
 export * from "./postLogout";
+export * from "./putChangePassword";
+export * from "./putSetPassword";
